:recycle: Extract walkToRectangle
